+++ src/SME.SGP.Dados/Repositorios/RepositorioNotificacaoConsulta.cs
@@ -239,34 +239,6 @@ namespace SME.SGP.Dados.Repositorios
             return await database.Conexao.QueryFirstAsync<int>(sql, new { anoLetivo, codigoRf, naoLida = (int)NotificacaoStatus.Pendente });
         }
 
-        public async Task<IEnumerable<NotificacaoBasicaDto>> ObterNotificacoesPorAnoLetivoERfAsync(int anoLetivo, string usuarioRf, int limite = 5)
-        {
-            var sql = @"select
-	                        n.id,
-	                        n.categoria,
-	                        n.codigo ,
-	                        n.criado_em as Data,
-	                        n.mensagem as DescricaoStatus,
-	                        n.status,
-	                        n.tipo,
-	                        n.titulo
-                        from
-	                        notificacao n
-                        left join usuario u on
-	                        n.usuario_id = u.id
-                        where
-	                        u.rf_codigo = @usuarioRf
-	                        and extract(year
-                        from
-	                        n.criado_em) = @anoLetivo
-	                        and not excluida
-                        order by
-	                        n.status asc,
-	                        n.criado_em desc
-                        limit @limite";
-            return await database.Conexao.QueryAsync<NotificacaoBasicaDto>(sql, new { anoLetivo, usuarioRf, limite });
-        }
-
         public async Task<IEnumerable<NotificacoesParaTratamentoCargosNiveisDto>> ObterNotificacoesParaTratamentoCargosNiveis()
         {
             var query = @"select
