+++ src/SME.SGP.Dominio.Interfaces/Repositorios/IRepositorioNotificacao.cs
@@ -1,3 +1,5 @@
+using SME.SGP.Infra;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace SME.SGP.Dominio.Interfaces
@@ -7,5 +9,6 @@ namespace SME.SGP.Dominio.Interfaces
         Task ExcluirPorIdsAsync(long[] ids);
         Task ExcluirLogicamentePorIdsAsync(long[] ids);
         Task ExcluirPeloSistemaAsync(long[] ids);
+        Task<IEnumerable<NotificacaoBasicaDto>> ObterNotificacoesPorAnoLetivoERfAsync(int anoLetivo, string usuarioRf, int limite = 5);
     }
 }
